# [Enhancement] Optimize append_selective for binary column 