## Why I'm doing:

Optimize `BinaryColumnBase<T>::append_selective`

1. **cache src offsets**: buffer the i-th start offset and end offset in `new_offsets[i * 2]` and `new_offsets[i * 2 + 1]`. This way, only one random access to `src_offset[indexes[i]]` is needed.  
2. **memcpy_inlined_overflow16**： For `src_column` bytes ≤ 32MB, use `memcpy_inlined_overflow16` instead of `memcpy_inlined`.  
3. **prefetch**: For `src_column` > 32MB, apply prefetching.  

str_len | num_src_rows | baseline (ns) | cache src offset (ns) | cache src offset + memcpy_inlined_overflow16 (ns) | cache src offset + prefetch (ns)
-- | -- | -- | -- | -- | --
16~32 | 4096 | 7730 | 5058 | 3744 | -
16~32 | 40960 | 54998 | 50926 | 36238 | -
16~32 | 409600 | 696769 | 610758 | 510106 | -
16~32 | 32MB | 165557080 | 126796247 | - | 85988980
4~8 | 4096 | 6600 | 6093 | 4627 | -
4~8 | 40960 | 65818 | 54135 | 41275 | -
4~8 | 409600 | 1168200 | 891207 | 687830 | -
4~8 | 32MB | 116201189 | 99602161 | - | 76007723


## What I'm doing:

Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 4.0
  - [ ] 3.5
  - [ ] 3.4
  - [ ] 3.3